fixed backend_test_utility openapi key display
