+++ backend_services/lib/backend_services_exports.dart
@@ -3,3 +3,4 @@ export 'package:backend_services/model/conversation.dart';
 export 'package:backend_services/src/state-management/conversations_provider.dart';
 export 'package:backend_services/src/state-management/sorting_type.dart';
 export 'package:backend_services/src/test-data/test_conversations.dart';
+export 'package:backend_services/src/environment_vars.dart';
