+++ backend_test_utility/lib/main.dart
@@ -1,6 +1,6 @@
+import 'package:backend_services/backend_services_exports.dart';
 import 'package:backend_test_utility/ambients.dart';
 import 'package:backend_test_utility/app_startup.dart';
-import 'package:backend_services/environment_vars.dart';
 import 'package:backend_test_utility/test_screens/app_instance_screen.dart';
 import 'package:backend_test_utility/test_screens/gpt_summary_screen.dart';
 import 'package:backend_test_utility/test_screens/websocket_screen.dart';
